narrow types of k and m from unsigned to unsigned short
Ignore-this: 4c26811e89232bb7917c961900670a87

darcs-hash:9bc59ddbad635e5bde000497cc81a50d75d0368c

+++ zfec/zfec/fec.c
@@ -412,7 +412,7 @@ fec_free (fec_t *p) {
 }
 
 fec_t *
-fec_new(unsigned k, unsigned n) {
+fec_new(unsigned short k, unsigned short n) {
     unsigned row, col;
     gf *p, *tmp_m;
